Add integration tests for tmf634 server
